feat(hot_reload.py): add hot reloading functionality for bot extensions
This commit introduces a new feature that allows for hot reloading of bot extensions. This is achieved by continuously checking for changes in extension files and reloading them if any modifications are detected. This feature improves the development workflow by eliminating the need to manually restart the bot whenever an extension is updated.