chore(workflow): update linting workflow to use Ruff for linting and formatting
Rename the workflow to "Ruff - Linting and Formatting" to reflect the
use of Ruff. Change job name from "Linting" to "Ruff" for clarity.
Remove the use of a custom token for checkout, relying on the default
GitHub token instead. Update the commit message to "chore" to better
align with the nature of the changes. These changes improve the
workflow's clarity and maintainability by clearly indicating the
linting tool and simplifying token usage.