style(influxdblogger.py): change log level from error to warning for InfluxDB init failure
Change the log level from error to warning when the InfluxDB logger
fails to initialize. This adjustment reflects that the failure to
initialize the InfluxDB logger is not necessarily a critical error,
but rather a warning that the feature will not be available unless
properly configured. This change helps to reduce unnecessary alarm
for users who may not require InfluxDB logging.