chore(Dockerfile): remove tealdeer package and update TLDR cache handling
Remove the tealdeer package from the Dockerfile as it is no longer
required. Instead, create a TLDR cache directory with appropriate
permissions for the nonroot user. This change ensures that the
application image is leaner by removing unnecessary packages and
improves security by setting up the cache directory with the correct
permissions for the nonroot user.