style(app.py): update logger warning messages for clarity and conciseness
Improve the clarity and conciseness of logger warning messages
regarding the use of eval commands by sysadmins. The updated messages
use emojis to quickly convey the nature of the warning and direct users
to the settings.yml.example file for more detailed information. This
change aims to make the warnings more user-friendly and informative.