ci(docker-test.yml): update permissions and handle errors in performance results
Add write permissions for issues and pull-requests to enable the workflow
to interact with GitHub issues and pull requests. This is necessary for
automating comments or updates related to the workflow's execution.

Set `continue-on-error: true` for the step that comments performance
results on pull requests. This ensures that the workflow continues even
if there is an error in this step, preventing the entire workflow from
failing due to issues in posting comments.